[Earnin Patch] Route RUM events to log files
This is a patch to route RUM events to log files instead of the datadog sdk sending it to its backend. The RUM events are cached and an API is provided for force flushing any cached events.

This patch should be applied to a forked RUM SDK which would only be used during performance testing. Whenever the datadog SDK version is updated, the forked SDK can be updated and this patch can be applied to the updated fork.

Target state is to open a feature request with Datadog and add this as a feature.